refactor(interpreter): split activation registration tests
